Add hvname to gethvs output + Initial work on getstats
Added the hypervariables name to the output of gethv, which is
necessary in order for getstats to find specific hypervariables.
Also started work on getstats, which currently reads results/hvs
outputted by gethv, reads supervariable definitions file, and is
in the middle of extracting said supervariables from hypervariables
outputted by gethv.